fix(spurdbd): cast SUM aggregates to BIGINT in get_usage query
PostgreSQL SUM() on BIGINT columns returns NUMERIC, causing a runtime
panic when sqlx tries to decode into i64.

+++ crates/spurdbd/src/db.rs
@@ -361,8 +361,10 @@ pub async fn get_usage(
 ) -> anyhow::Result<Vec<UsageRecord>> {
     let rows = sqlx::query(
         r#"
-        SELECT user_name, account, SUM(cpu_seconds) as total_cpu_seconds,
-               SUM(gpu_seconds) as total_gpu_seconds, SUM(job_count) as total_jobs,
+        SELECT user_name, account,
+               SUM(cpu_seconds)::BIGINT as total_cpu_seconds,
+               SUM(gpu_seconds)::BIGINT as total_gpu_seconds,
+               SUM(job_count)::BIGINT as total_jobs,
                period_start
         FROM usage
         WHERE period_start >= $1
